fix: fix problems in review
