+++ c-pallets/cacher/src/lib.rs
@@ -8,6 +8,9 @@ pub mod mock;
 mod tests;
 pub mod weights;
 
+mod types;
+use types::*;
+
 use frame_system::pallet_prelude::*;
 use frame_support::{
 	pallet_prelude::*,
@@ -15,39 +18,13 @@ use frame_support::{
 		Currency, LockableCurrency,
 		ExistenceRequirement::KeepAlive,
 	},
+	transactional,
 };
-use cp_cess_common::{
-	IpAddress,
-};
+use cp_cess_common::IpAddress;
 
 pub use pallet::*;
-use sp_runtime::traits::Zero;
 use sp_std::prelude::*;
 pub use weights::WeightInfo;
-
-type AccountOf<T> = <T as frame_system::Config>::AccountId;
-/// The balance type of this pallet.
-pub type BalanceOf<T> =
-	<<T as Config>::Currency as Currency<<T as frame_system::Config>::AccountId>>::Balance;
-
-/// The custom struct for cacher info.
-#[derive(PartialEq, Eq, Encode, Decode, Clone, RuntimeDebug, MaxEncodedLen, TypeInfo)]
-pub struct CacherInfo<AccoutId, Balance> {
-	pub acc: AccoutId,
-	pub ip: IpAddress,
-	pub byte_price: Balance,
-}
-
-/// The custom struct for bill info.
-#[derive(PartialEq, Eq, Encode, Decode, Clone, RuntimeDebug, MaxEncodedLen, TypeInfo)]
-pub struct Bill<AccoutId, Balance, Hash> {
-	pub id: [u8; 16], 
-	pub to: AccoutId,
-	pub amount: Balance,
-	pub file_hash: Hash,
-	pub slice_hash: Hash,
-	pub expiration_time: u64,
-}
 
 #[frame_support::pallet]
 pub mod pallet {
@@ -80,19 +57,17 @@ pub mod pallet {
 	#[pallet::error]
 	pub enum Error<T> {
 		//Registered Error
-		Registered,
+		AlreadyRegistered,
 		//Unregistered Error
-		UnRegister,
+		UnRegistered,
 		//Option parse Error
 		OptionParseError,
-		//Insufficient balance Error
-		InsufficientBalance,
 	}
 
 	/// Store all cacher info
 	#[pallet::storage]
 	#[pallet::getter(fn cacher)]
-	pub(super) type Cachers<T: Config> = StorageMap<_, Twox64Concat, AccountOf<T>, CacherInfo<AccountOf<T>, BalanceOf<T>>>;
+	pub(super) type Cachers<T: Config> = StorageMap<_, Blake2_128Concat, AccountOf<T>, CacherInfo<AccountOf<T>, BalanceOf<T>>>;
 
 	#[pallet::pallet]
 	#[pallet::generate_store(pub(super) trait Store)]
@@ -108,7 +83,7 @@ pub mod pallet {
 		#[pallet::weight(T::WeightInfo::register())]
 		pub fn register(origin: OriginFor<T>, info: CacherInfo<AccountOf<T>, BalanceOf<T>>) -> DispatchResult {
 			let sender = ensure_signed(origin)?;
-			ensure!(!<Cachers<T>>::contains_key(&sender), Error::<T>::Registered);
+			ensure!(!<Cachers<T>>::contains_key(&sender), Error::<T>::AlreadyRegistered);
 			<Cachers<T>>::insert(&sender, info.clone());
 
 			Self::deposit_event(Event::<T>::Register {acc: sender, info});
@@ -123,7 +98,7 @@ pub mod pallet {
 		#[pallet::weight(T::WeightInfo::update())]
 		pub fn update(origin: OriginFor<T>, info: CacherInfo<AccountOf<T>, BalanceOf<T>>) -> DispatchResult {
 			let sender = ensure_signed(origin)?;
-			ensure!(<Cachers<T>>::contains_key(&sender), Error::<T>::UnRegister);
+			ensure!(<Cachers<T>>::contains_key(&sender), Error::<T>::UnRegistered);
 
 			<Cachers<T>>::try_mutate(&sender, |info_opt| -> DispatchResult {
 				let p_info = info_opt.as_mut().ok_or(Error::<T>::OptionParseError)?;
@@ -140,7 +115,7 @@ pub mod pallet {
 		#[pallet::weight(T::WeightInfo::logout())]
 		pub fn logout(origin: OriginFor<T>) -> DispatchResult {
 			let sender = ensure_signed(origin)?;
-			ensure!(<Cachers<T>>::contains_key(&sender), Error::<T>::UnRegister);
+			ensure!(<Cachers<T>>::contains_key(&sender), Error::<T>::UnRegistered);
 
 			<Cachers<T>>::remove(&sender);
 
@@ -153,16 +128,12 @@ pub mod pallet {
 		///	
 		/// Parameters:
 		/// - `bills`: list of bill.
+		#[transactional]
 		#[pallet::weight(T::WeightInfo::pay(bills.len() as u32))]
 		pub fn pay(origin: OriginFor<T>, bills: Vec<Bill<AccountOf<T>, BalanceOf<T>, T::Hash>>) -> DispatchResult {
 			let sender = ensure_signed(origin)?;
-			let mut total_amount: BalanceOf<T> = Zero::zero();
-			for bill in bills.iter() {
-				total_amount += bill.amount;
-			}
-			ensure!(T::Currency::free_balance(&sender) >= total_amount, Error::<T>::InsufficientBalance);
 			
-			for bill in bills.iter() {
+			for bill in &bills {
 				T::Currency::transfer(&sender, &bill.to, bill.amount, KeepAlive)?;
 			}
 			
@@ -172,4 +143,3 @@ pub mod pallet {
 		}
 	}
 }
-
